ai-inference-backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
websockets==14.2
tensorflow==2.12.0
numpy==2.0.2
pillow==11.1.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
requests==2.32.3
# opencv-python==4.11.0.86  # only if you use OpenCV somewhere (optional)
